refactor: canonicalize ContainsShr and CollectVariables in ExprUtils
Three near-identical Shr-detection predicates (DynamicMask::ContainsShr,
StructureRecovery::HasShr, SemilinearNormalizer::ContainsShr) and two
near-identical variable-collection helpers (SemilinearIR::CollectVarsFromExpr,
SemilinearNormalizer::CollectVariables) collapse into a single ExprUtils pair.

Closes audit findings semilinear-decomposition-cross-3, -6, and -5.
